pyproject.toml
[project]
name = "assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "neo4j>=5.28.2",
    "openai>=1.108.1",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
